# swagger_output.json
{"openapi":"3.1.0","info":{"title":"Maejang API","description":"<h2>배달 사이트 프로젝트</h2> API 명세서","version":"1.0"},"servers":[{"url":"http://3.24.199.215:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/users/sign_in":{"post":{"tags":["user-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseUserIdResponse"}}}}}}},"/api/v1/users/sing_in":{"post":{"tags":["user-controller"],"operationId":"signup_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseUserIdResponse"}}}}}}},"/api/v1/users/logout":{"post":{"tags":["user-controller"],"operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/users/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseUserIdResponse"}}}}}}},"/api/v1/users/email/duplicate":{"post":{"tags":["user-controller"],"operationId":"emailDuplicate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDuplicateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseEmailDuplicateResponse"}}}}}}},"/api/v1/store/open":{"post":{"tags":["store-controller"],"summary":"가게 오픈","description":"현재 로그인한 사장의 가게를 오픈 처리합니다.","operationId":"open","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/store/create":{"post":{"tags":["store-controller"],"summary":"가게 생성","description":"사장당 1개의 가게를 생성합니다.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseStoreIdResponse"}}}}}}},"/api/v1/store/close":{"post":{"tags":["store-controller"],"summary":"가게 클로즈","description":"현재 로그인한 사장의 가게를 클로즈 처리합니다.","operationId":"close","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/order/ok":{"post":{"tags":["order-controller"],"summary":"주문수락","description":"사장이 주문을 수락합니다.","operationId":"ok","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/order/create":{"post":{"tags":["order-controller"],"summary":"주문하기","description":"주문을 생성합니다.","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseOrderIdResponse"}}}}}}},"/api/v1/order/cancel":{"post":{"tags":["order-controller"],"summary":"주문거절","description":"사장이 주문을 거절합니다.","operationId":"reject","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/menu/create":{"post":{"tags":["menu-controller"],"summary":"메뉴 신규 등록","description":"메뉴를 신규 등록합니다.","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuUpsertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseMenuIdResponse"}}}}}}},"/api/v1/cart/create":{"post":{"tags":["cart-controller"],"summary":"장바구니 등록","description":"장바구니에 메뉴를 추가합니다.","operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseCartItemIdResponse"}}}}}}},"/api/v1/auth/me":{"post":{"tags":["auth-controller"],"summary":"내 정보","description":"JWT 쿠키 기반 인증이 정상인지 확인용","operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseAuthMeResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"summary":"로그아웃(쿠키 삭제)","description":"ACCESS_TOKEN 쿠키를 만료시킵니다.","operationId":"logout_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"summary":"로그인","description":"응답 헤더 Authorization: Bearer <token> 으로 토큰을 내려줍니다. (Swagger에서 복사해서 Authorize에 붙여넣어 사용)","operationId":"login_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/address/create":{"post":{"tags":["address-controller"],"summary":"주소 신규 등록","description":"주소를 신규 등록합니다.","operationId":"create_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseAddressIdResponse"}}}}}}},"/api/v1/address/choose/{addressId}":{"post":{"tags":["address-controller"],"summary":"주소 선택","description":"주소를 선택합니다.","operationId":"choose","parameters":[{"name":"addressId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/menu/update/{menuId}":{"patch":{"tags":["menu-controller"],"summary":"메뉴 수정","description":"메뉴를 수정합니다.","operationId":"update","parameters":[{"name":"menuId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/address/update/{addressId}":{"patch":{"tags":["address-controller"],"summary":"주소 수정","description":"주소를 수정합니다.","operationId":"update_1","parameters":[{"name":"addressId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/store/read":{"get":{"tags":["store-controller"],"summary":"가게 조회","description":"현재 로그인한 사장의 가게를 조회합니다.","operationId":"read","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseStoreResponse"}}}}}}},"/api/v1/order/read":{"get":{"tags":["order-controller"],"summary":"주문 확인(1건)","description":"주문 1건을 조회합니다.","operationId":"read_1","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseOrder"}}}}}}},"/api/v1/order/history":{"get":{"tags":["order-controller"],"summary":"과거 주문 내역","description":"사용자의 주문 내역을 조회합니다.","operationId":"history","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseListOrder"}}}}}}},"/api/v1/order/check":{"get":{"tags":["order-controller"],"summary":"사장 주문조회","description":"TODO: 사장 주문조회(가게 기준) 구현 전 임시","operationId":"check","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/menu/read":{"get":{"tags":["menu-controller"],"summary":"메뉴 목록 조회","description":"메뉴 목록을 조회합니다.","operationId":"read_2","parameters":[{"name":"ownerId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseListMenuResponse"}}}}}}},"/api/v1/cart/read":{"get":{"tags":["cart-controller"],"summary":"장바구니 조회","description":"사용자의 장바구니 목록을 조회합니다.","operationId":"read_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseListCartItemResponse"}}}}}}},"/api/v1/address/read":{"get":{"tags":["address-controller"],"summary":"주소 목록 조회","description":"주소 목록을 조회합니다.","operationId":"read_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseListAddressResponse"}}}}}}},"/api/v1/order/delete":{"delete":{"tags":["order-controller"],"summary":"주문취소","description":"고객이 주문을 취소합니다.","operationId":"cancel","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/menu/delete/{menuId}":{"delete":{"tags":["menu-controller"],"summary":"메뉴 삭제","description":"메뉴를 삭제합니다.","operationId":"delete","parameters":[{"name":"menuId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/cart/delete/{cartItemId}":{"delete":{"tags":["cart-controller"],"summary":"장바구니 삭제","description":"장바구니 항목을 삭제합니다.","operationId":"delete_1","parameters":[{"name":"cartItemId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}},"/api/v1/address/delete/{addressId}":{"delete":{"tags":["address-controller"],"summary":"주소 삭제","description":"주소를 삭제합니다.","operationId":"delete_2","parameters":[{"name":"addressId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JSONResponseVoid"}}}}}}}},"components":{"schemas":{"UserSignupRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["OWNER","CUSTOMER"]},"name":{"type":"string"}},"required":["email","name","password","role"]},"JSONResponseUserIdResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserIdResponse"}}},"UserIdResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"}}},"JSONResponseVoid":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"object"}}},"UserLoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"EmailDuplicateRequest":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"EmailDuplicateResponse":{"type":"object","properties":{"duplicated":{"type":"boolean"}}},"JSONResponseEmailDuplicateResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/EmailDuplicateResponse"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"StoreCreateRequest":{"type":"object","properties":{"storeName":{"type":"string"},"address":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"},"openTime":{"$ref":"#/components/schemas/LocalTime"},"closeTime":{"$ref":"#/components/schemas/LocalTime"}},"required":["address","storeName"]},"JSONResponseStoreIdResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/StoreIdResponse"}}},"StoreIdResponse":{"type":"object","properties":{"storeId":{"type":"integer","format":"int64"}}},"OrderCreateRequest":{"type":"object","properties":{"request":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemRequest"}}},"required":["items"]},"OrderItemRequest":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"},"option":{"type":"string"},"count":{"type":"integer","format":"int32"}},"required":["count","menuId"]},"JSONResponseOrderIdResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/OrderIdResponse"}}},"OrderIdResponse":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"}}},"MenuUpsertRequest":{"type":"object","properties":{"menuName":{"type":"string"},"picture":{"type":"string"},"price":{"type":"integer","format":"int32","minimum":0},"description":{"type":"string"},"option":{"type":"string"}},"required":["menuName"]},"JSONResponseMenuIdResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/MenuIdResponse"}}},"MenuIdResponse":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"}}},"CartCreateRequest":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"},"option":{"type":"string"},"count":{"type":"integer","format":"int32","minimum":1}},"required":["menuId"]},"CartItemIdResponse":{"type":"object","properties":{"cartItemId":{"type":"integer","format":"int64"}}},"JSONResponseCartItemIdResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/CartItemIdResponse"}}},"AuthMeResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"email":{"type":"string"}}},"JSONResponseAuthMeResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthMeResponse"}}},"AddressCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"}},"required":["address","name"]},"AddressIdResponse":{"type":"object","properties":{"addressId":{"type":"integer","format":"int64"},"name":{"type":"string"},"address":{"type":"string"}}},"JSONResponseAddressIdResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AddressIdResponse"}}},"MenuUpdateRequest":{"type":"object","properties":{"menuName":{"type":"string"},"picture":{"type":"string"},"price":{"type":"integer","format":"int32"},"description":{"type":"string"},"option":{"type":"string"}}},"AddressUpdateRequest":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"}}},"JSONResponseStoreResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/StoreResponse"}}},"StoreResponse":{"type":"object","properties":{"storeId":{"type":"integer","format":"int64"},"ownerId":{"type":"integer","format":"int64"},"storeName":{"type":"string"},"address":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"},"openTime":{"$ref":"#/components/schemas/LocalTime"},"closeTime":{"$ref":"#/components/schemas/LocalTime"},"isOpen":{"type":"boolean"}}},"JSONResponseOrder":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/Order"}}},"Order":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"price":{"type":"integer","format":"int32"},"request":{"type":"string"},"condition":{"type":"string","enum":["ORDERED","COOKING","DELIVERING","DELIVERED","REJECTED","CANCELLED"]},"orderAt":{"type":"string","format":"date-time"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["OWNER","CUSTOMER"]},"name":{"type":"string"}}},"JSONResponseListOrder":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}},"JSONResponseListMenuResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/MenuResponse"}}}},"MenuResponse":{"type":"object","properties":{"menuId":{"type":"integer","format":"int64"},"ownerId":{"type":"integer","format":"int64"},"menuName":{"type":"string"},"picture":{"type":"string"},"price":{"type":"integer","format":"int32"},"description":{"type":"string"},"option":{"type":"string"}}},"CartItemResponse":{"type":"object","properties":{"cartItemId":{"type":"integer","format":"int64"},"menuId":{"type":"integer","format":"int64"},"menuName":{"type":"string"},"option":{"type":"string"},"count":{"type":"integer","format":"int32"}}},"JSONResponseListCartItemResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CartItemResponse"}}}},"AddressResponse":{"type":"object","properties":{"addressId":{"type":"integer","format":"int64"},"name":{"type":"string"},"address":{"type":"string"},"isDefault":{"type":"boolean"}}},"JSONResponseListAddressResponse":{"type":"object","description":"결과 공통 정보","properties":{"success":{"type":"boolean"},"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/AddressResponse"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}